Move the attestation handling code to separate file for reuse
